fix(report): keep the submit error visible beside the pinned action
Pinning the submit action let the user submit from any scroll offset, but
the inline error still rendered at the end of the scroll content — below
eleven reason cards. On a 412x915 device, submitting from the top with
"Spam" selected put the error at y=1692, roughly 780px past the fold: the
spinner stopped and nothing else appeared to happen. In the pre-pinning
layout the in-flow button forced the user to the bottom, so the error was
always adjacent to it.

The error now rides the footer directly above the button, so the two stay
together at any scroll offset. The regression test pins this at phone size
by asserting the error's rect falls inside the viewport; it fails at
1708.25 > 915.0 without the fix.

Along the way:

- Promote the sheet's private _KeyboardAwareBottomInput to a shared
  VineKeyboardAwareFooter in divine_ui rather than keeping a second copy
  of the same clearance and animation constants in the report sheet.
- Split the confirmation state into report_content_confirmation.dart.
  That takes the dialog from 886 to 703 lines, back under the advisory
  800-line ceiling, so its stale baseline entry can go rather than be
  raised.
- Drop a tester.ensureVisible on the submit button that became a no-op
  once the footer moved outside the scroll view.
