+++ mobile/packages/divine_ui/lib/src/bottom_sheet/bottom_sheet.dart
@@ -6,3 +6,4 @@ export 'vine_bottom_sheet_drag_handle.dart';
 export 'vine_bottom_sheet_header.dart';
 export 'vine_bottom_sheet_prompt.dart';
 export 'vine_bottom_sheet_selection_menu.dart';
+export 'vine_keyboard_aware_footer.dart';
